Make use of the MMU during symbolic execution optional
Running without the MMU is much faster.

This relies on the chosen instructions in each test being ambivalent
about whether address translation is actually on, at least in the Sail.
The generated test will still turn it on so that instructions like
exclusive load/stores can be tested - but the model doesn't care.